Fixed importing of math extension modules.
This change addresses three issues:

1. The extension modules in charm/core/math had a call to PyType_Ready
for BenchmarkType, which is already done in the benchmark module. This
caused the import of these modules to fail.

2. The integermodule had an inline function (size). By default, clang
uses C99 semantics for the inline keyword. This leads to a "Symbol
not found: _size" on macOS. The function was changed to static inline.

3. Modules charm/schemes/abenc/{ac17|bsw07|cgw15|waters11} import msp
instead of charm.schemes.abenc.msp.